salesforce_ingest/pyproject.toml
# pyproject.toml
# Centralized tool configuration for the salesforce_ingest project.
# All tools (pytest, ruff, mypy) read from this file automatically.

[build-system]
requires      = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name        = "salesforce_ingest"
version     = "1.0.0"
description = "Salesforce → S3 → Snowpipe → dbt incremental data pipeline"
requires-python = ">=3.12"

dependencies = [
    "requests>=2.32",
    "PyJWT>=2.9",
    "cryptography>=43",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-mock>=3.14",
    "pytest-cov>=5",
    "moto[s3,ssm,secretsmanager,core]>=5",
    "ruff>=0.6",
    "mypy>=1.11",
    "boto3-stubs[ssm,s3,secretsmanager,lambda,stepfunctions,codebuild]>=1.34",
    "snowflake-connector-python>=3.12",
]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths       = ["tests"]
python_files    = ["test_*.py"]
python_classes  = ["Test*"]
python_functions= ["test_*"]
addopts         = [
    "--tb=short",
    "--strict-markers",
    "--cov=lambda",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=70",
]
markers = [
    "unit:        Fast unit tests with mocked AWS (no real cloud resources)",
    "integration: Slow integration tests requiring live AWS credentials",
    "snowflake:   Tests requiring a live Snowflake connection",
]
filterwarnings = [
    "ignore::DeprecationWarning:botocore",
    "ignore::DeprecationWarning:urllib3",
]

# ---------------------------------------------------------------------------
# ruff (linter + formatter)
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length    = 100
src            = ["lambda", "glue", "scripts", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "LOG",  # flake8-logging
    "G",    # flake8-logging-format
    "PIE",  # flake8-pie
    "RET",  # flake8-return
]
ignore = [
    "E501",   # line too long — handled by formatter
    "B008",   # do not perform function calls in default arguments
    "G004",   # logging-f-string — we intentionally use f-strings for readability
    "RET504", # unnecessary assignment before return
    "SIM108", # ternary operator — sometimes explicit is clearer
]

[tool.ruff.lint.isort]
known-first-party = ["salesforce_client", "watermark", "handler"]
force-sort-within-sections = true

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]   # assert is fine in tests

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version          = "3.12"
strict                  = false
warn_return_any         = true
warn_unused_configs     = true
warn_redundant_casts    = true
warn_unused_ignores     = true
disallow_untyped_defs   = true
disallow_incomplete_defs= true
check_untyped_defs      = true
no_implicit_optional    = true
show_error_codes        = true
pretty                  = true

[[tool.mypy.overrides]]
module = [
    "snowflake.*",
    "moto.*",
    "cryptography.*",
    "jwt.*",
    "awsglue.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source   = ["lambda"]
omit     = ["*/__init__.py", "*/test_*", "*/conftest.py"]
branch   = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
show_missing = true
